Unify set and lists to lists in constnats, move all constnat into a separate submodule
