+++ mung/constants/__init__.py
@@ -1,0 +1,9 @@
+"""This module implements constants that are used inside the pitch,
+duration and onset inference algorithm."""
+
+
+from .general import WESTERN_NOTATION_STAFFLINE_COUNT, STEP_ORDER
+from .onsets import OnsetDataConstants
+from .precedence_links import PrecedenceLinksConstants
+from .class_names import ClassNamesConstants
+from .inference import InferenceEngineConstants
