+++ mung/constants/onsets.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class OnsetDataConstants(Enum):
+    ONSET_BEATS = "onset_beats"
+    DURATION_BEATS = "duration_beats"
+    DURATION_BEATS_WO_M = "duration_beats_wo_m"
+    TUPLE_TIME_MODIFICATION = "time_modification"
