+++ mung/constants/precedence_links.py
@@ -1,0 +1,6 @@
+class PrecedenceLinksConstants:
+    """
+    This class stores names of precedence-link-related fields in ``Node.data``.
+    """
+    PrecedenceInlinks: str = "precedence_inlinks"
+    PrecedenceOutlinks: str = "precedence_outlinks"
